+++ djangocms_blog/migrations/0004_auto_20150108_1435.py
@@ -4,6 +4,7 @@ from __future__ import unicode_literals
 import django.db.models.deletion
 import filer.fields.image
 from django.db import migrations, models
+from djangocms_blog.models import thumbnail_model
 from filer.settings import FILER_IMAGE_MODEL
 
 ACTUAL_FILER_IMAGE_MODEL = FILER_IMAGE_MODEL or 'filer.Image'
@@ -26,13 +27,13 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='post',
             name='main_image_full',
-            field=models.ForeignKey(related_name='djangocms_blog_post_full', on_delete=django.db.models.deletion.SET_NULL, verbose_name='Main image full', blank=True, to='cmsplugin_filer_image.ThumbnailOption', null=True),
+            field=models.ForeignKey(related_name='djangocms_blog_post_full', on_delete=django.db.models.deletion.SET_NULL, verbose_name='Main image full', blank=True, to=thumbnail_model, null=True),
             preserve_default=True,
         ),
         migrations.AlterField(
             model_name='post',
             name='main_image_thumbnail',
-            field=models.ForeignKey(related_name='djangocms_blog_post_thumbnail', on_delete=django.db.models.deletion.SET_NULL, verbose_name='Main image thumbnail', blank=True, to='cmsplugin_filer_image.ThumbnailOption', null=True),
+            field=models.ForeignKey(related_name='djangocms_blog_post_thumbnail', on_delete=django.db.models.deletion.SET_NULL, verbose_name='Main image thumbnail', blank=True, to=thumbnail_model, null=True),
             preserve_default=True,
         ),
     ]