+++ djangocms_blog/models.py
@@ -26,6 +26,13 @@ from .settings import get_setting
 BLOG_CURRENT_POST_IDENTIFIER = get_setting('CURRENT_POST_IDENTIFIER')
 BLOG_CURRENT_NAMESPACE = get_setting('CURRENT_NAMESPACE')
 
+try:
+    from filer.models import ThumbnailOption  # NOQA
+    thumbnail_model = 'filer.ThumbnailOption'
+except ImportError:
+    from cmsplugin_filer_image.models import ThumbnailOption  # NOQA
+    thumbnail_model = 'cmsplugin_filer_image.ThumbnailOption'
+
 
 @python_2_unicode_compatible
 class BlogCategory(TranslatableModel):
@@ -104,12 +111,12 @@ class Post(ModelMeta, TranslatableModel):
     main_image = FilerImageField(verbose_name=_('main image'), blank=True, null=True,
                                  on_delete=models.SET_NULL,
                                  related_name='djangocms_blog_post_image')
-    main_image_thumbnail = models.ForeignKey('cmsplugin_filer_image.ThumbnailOption',
+    main_image_thumbnail = models.ForeignKey(thumbnail_model,
                                              verbose_name=_('main image thumbnail'),
                                              related_name='djangocms_blog_post_thumbnail',
                                              on_delete=models.SET_NULL,
                                              blank=True, null=True)
-    main_image_full = models.ForeignKey('cmsplugin_filer_image.ThumbnailOption',
+    main_image_full = models.ForeignKey(thumbnail_model,
                                         verbose_name=_('main image full'),
                                         related_name='djangocms_blog_post_full',
                                         on_delete=models.SET_NULL,